My
This repository is created for saving my Computer Science Practicals programs to access it anywhere. I have coded most of the programs that is from the 1st program to 15th program. In that 14th and the 15th program is not working as intended, Iam currently working on a solution for it. Program 16 to the end is not required. If at-all you have any queries feel free to raise and issue. Thank You! ~JuliaNGaminG